deployment/requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib
plotly
Pillow
